refactor(library/utilitiesByType): moves "/array.ts" to "/array/hasAtLeastOne.ts"
